voir les candidatures d'une session-formation donnée

+++ app/Http/Controllers/SessionController.php
@@ -15,29 +15,37 @@ class SessionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function index()
-    {   abort_if(Gate::denies('session_index'), 403);
+    {
+        abort_if(Gate::denies('session_index'), 403);
         $sessions = Session::all()->sortBy('date_session')->sortByDesc('annee_univ');
         return view('admin.session.index', compact('sessions'));
     }
 
+    public function all(Request $req)
+    {
+        $sessions = Session::all()->sortByDesc('date_session');
+
+        return view('admin.candidature.sessions')->with('sessions', $sessions);
+    }
     /**
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
      */
     public function create()
-    {    abort_if(Gate::denies('session_create'), 403);
+    {
+        abort_if(Gate::denies('session_create'), 403);
         return view('admin.session.create');
     }
 
-     function listeAnneeUniv()
-    {   
-        $t[0] = strval(date('Y')."-".(date('Y')+1));
-        for ($i = 0; $i <20; $i++){
+    function listeAnneeUniv()
+    {
+        $t[0] = strval(date('Y') . "-" . (date('Y') + 1));
+        for ($i = 0; $i < 20; $i++) {
             $currentYear = date('Y');
             $year = $currentYear - $i;
-            $lastYear = $currentYear-($i+1);
-            $t[$i+1] = strval($lastYear."-".$year);
+            $lastYear = $currentYear - ($i + 1);
+            $t[$i + 1] = strval($lastYear . "-" . $year);
         }
         return $t;
     }
@@ -49,12 +57,13 @@ class SessionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {   abort_if(Gate::denies('session_create'), 403);
+    {
+        abort_if(Gate::denies('session_create'), 403);
         $t = $this->listeAnneeUniv();
-        $request->session()->flash('operation','store');
+        $request->session()->flash('operation', 'store');
         $request->validate([
-            'annee_univ' => 'required|in:'.implode(',',$t),
-            'date_session'=>['required', 'date_format:Y-m-d']
+            'annee_univ' => 'required|in:' . implode(',', $t),
+            'date_session' => ['required', 'date_format:Y-m-d']
         ]);
 
         Session::create([
@@ -63,7 +72,7 @@ class SessionController extends Controller
         ]);
 
         return redirect()->route('session.index')
-         ->with('success','La session a été enregistrée') ;
+            ->with('success', 'La session a été enregistrée');
     }
 
     /**
@@ -73,11 +82,12 @@ class SessionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
-    {   abort_if(Gate::denies('session_edit'), 403);
+    {
+        abort_if(Gate::denies('session_edit'), 403);
         $session = Session::findOrFail($id);
         $response = [
             'session' => $session,
-            'route' =>  route('session.update',[$id])
+            'route' =>  route('session.update', [$id])
         ];
         return response()->json($response, 200);
     }
@@ -90,21 +100,21 @@ class SessionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
-    {   abort_if(Gate::denies('session_edit'), 403);
+    {
+        abort_if(Gate::denies('session_edit'), 403);
         $t = $this->listeAnneeUniv();
-        $request->session()->flash('operation','update');
+        $request->session()->flash('operation', 'update');
         $request->validate([
-            'annee_univ' => 'required|in:'.implode(',',$t),
-            'date_session'=>['required', 'date_format:Y-m-d']
+            'annee_univ' => 'required|in:' . implode(',', $t),
+            'date_session' => ['required', 'date_format:Y-m-d']
         ]);
 
-        Session::where('id',$id)->update([
+        Session::where('id', $id)->update([
             'annee_univ' => $request->get('annee_univ'),
             'date_session' => $request->get('date_session')
         ]);
         return redirect()->route('session.index')
-        ->with('success','La session a été modifié');
-      
+            ->with('success', 'La session a été modifié');
     }
 
     /**
@@ -114,14 +124,16 @@ class SessionController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {    abort_if(Gate::denies('session_delete'), 403);
+    {
+        abort_if(Gate::denies('session_delete'), 403);
         $session = Session::findOrFail($id);
         $session->delete();
         return redirect()->route('session.index')
-        ->with('success','La session a été supprimée');
+            ->with('success', 'La session a été supprimée');
     }
 
-    public function renderSessions(){
+    public function renderSessions()
+    {
         $sessions = Session::all();
         return  response()->json($sessions, 200);
     }
